clang

+++ src/activity_actor.cpp
@@ -711,7 +711,8 @@ static int hack_level( const Character &who )
     // odds go up with int>8, down with int<8
     // 4 int stat is worth 1 computer skill here
     ///\EFFECT_INT increases success chance of hacking card readers
-    return round( who.get_skill_level( skill_computer ) ) + who.int_cur / 2 - 8;
+    return round( who.get_skill_level( skill_computer ) + static_cast<float>
+                  ( who.int_cur ) / 2.0f - 8 );
 }
 
 static hack_result hack_attempt( Character &who )
